refactor(renderer): keep aabbEdgeLineList internal (no new public export)
Move the clash-box edge helper into its own module instead of exporting it from
the package index, so it's not added to the public API surface. Update the
changeset to document Renderer.setClashOverlapBox.
