+++ .changeset/feat-clash-highlight-colors.md
@@ -2,4 +2,9 @@
 "@ifc-lite/renderer": minor
 ---
 
-Add `RenderOptions.highlightColors` — a per-element highlight tint for selected meshes. A selected mesh whose id is in this map glows in the given RGBA using the same re-lit selection treatment instead of the default selection blue. Lets a consumer render, e.g., the two sides of a clash in distinct vibrant colours while keeping the selection glow, crease structure and opacity. Only applied to ids that are also selected (`selectedId`/`selectedIds`); absent map = unchanged behaviour. (#1277/#1339)
+Clash visualisation support:
+
+- Add `RenderOptions.highlightColors` — a per-element highlight tint. An id in this map glows in the given RGBA using the re-lit selection treatment (and gets the same individual-mesh / opaque / stay-solid-through-ghost handling) **without** needing to be in the store selection. Lets a consumer render, e.g., the two sides of a clash in distinct vibrant colours in highlight/isolate/ghost with no "selected" state.
+- Add `Renderer.setClashOverlapBox(box | null)` — draws a world-space AABB as a distinct-colour wireframe box (e.g. the clash overlap region) via the existing overlay line pipeline. Pass `null` to clear.
+
+(#1277/#1339)
