+++ packages/renderer/src/aabb-edges.test.ts
@@ -5,7 +5,7 @@
 import { describe, it } from 'node:test';
 import assert from 'node:assert';
 
-import { aabbEdgeLineList } from './index.ts';
+import { aabbEdgeLineList } from './aabb-edges.ts';
 
 describe('aabbEdgeLineList (#1277 clash overlap box)', () => {
   it('emits 12 edges (24 vertices, 72 floats)', () => {
