+++ packages/renderer/src/index.ts
@@ -17,6 +17,7 @@ export { Picker } from './picker.js';
 export { MathUtils } from './math.js';
 export { SectionPlaneRenderer } from './section-plane.js';
 export { Section2DOverlayRenderer } from './section-2d-overlay.js';
+import { aabbEdgeLineList } from './aabb-edges.js';
 
 // IfcAnnotation overlay pipelines (3D world-space). Self-contained — caller
 // passes a GPUDevice + presentation format and invokes `.render(pass, viewProj)`
@@ -153,36 +154,6 @@ type ResolvedVisualEnhancement = {
  * share the same aggregate position-length total can't collide on the
  * same fingerprint and reuse a stale BVH.
  */
-/**
- * Build a flat line-list (`[x,y,z, …]`, 12 edges = 24 vertices) for the 12 edges
- * of a world-space AABB. Used to draw the clash-overlap wireframe box (#1277).
- */
-export function aabbEdgeLineList(
-    min: readonly [number, number, number],
-    max: readonly [number, number, number],
-): Float32Array {
-    const [x0, y0, z0] = min;
-    const [x1, y1, z1] = max;
-    // 8 corners
-    const c = [
-        [x0, y0, z0], [x1, y0, z0], [x1, y1, z0], [x0, y1, z0], // bottom (z0)
-        [x0, y0, z1], [x1, y0, z1], [x1, y1, z1], [x0, y1, z1], // top (z1)
-    ];
-    // 12 edges as corner-index pairs
-    const edges = [
-        [0, 1], [1, 2], [2, 3], [3, 0], // bottom ring
-        [4, 5], [5, 6], [6, 7], [7, 4], // top ring
-        [0, 4], [1, 5], [2, 6], [3, 7], // verticals
-    ];
-    const out = new Float32Array(edges.length * 2 * 3);
-    let o = 0;
-    for (const [a, b] of edges) {
-        out[o++] = c[a][0]; out[o++] = c[a][1]; out[o++] = c[a][2];
-        out[o++] = c[b][0]; out[o++] = c[b][1]; out[o++] = c[b][2];
-    }
-    return out;
-}
-
 function computeBvhFingerprint(meshes: ReadonlyArray<import('@ifc-lite/geometry').MeshData>): string {
     const parts: string[] = [String(meshes.length)];
     for (const m of meshes) {
